refactor: refine main.js
